Modularize chat SDK messenger example boundaries
Split the chat-sdk-messenger example along the boundaries we expect to matter for a future SDK extraction. Keep ChatIngressAgent focused on orchestration while moving provider policy, admin directory mapping, and AI reply delivery helpers into smaller modules.

The Telegram provider module now owns webhook setup, HTTPS validation, message delivery limits, long-message chunking, final-edit no-op classification, and Telegram-specific state key sharding. This removes Telegram-specific dedupe key handling from the generic Agents-backed Chat SDK state adapter.

The intelligence delivery module now owns managed AI reply fiber names, snapshots, recovery mode, and failure mode policy. The admin directory module owns conversation/job DTOs and row-to-view mapping helpers. These files make the example a better staging ground for deciding what eventually belongs in agents/src/chat-sdk without prematurely extracting Telegram or admin UI product code.

Replace the ChatStateAgent recurring cleanup interval with lazy one-shot cleanup scheduling. Expiring writes now schedule the next cleanup only when needed, update the scheduled cleanup when an earlier expiry appears, and reschedule after cleanup only if expiring rows remain. Cleanup metadata tracks the next expiry and schedule id so state shards without expiring records do not pay for a forever schedule, which is a better shape for SDK-owned state infrastructure.

Update README/TODO notes to document the module layout and extraction candidates, and adjust tests to cover provider-owned Telegram key sharding instead of baking Telegram behavior into the generic state adapter.
