+++ examples/chat-sdk-messenger/src/index.ts
@@ -3,14 +3,34 @@ import { Agent, callable, getAgentByName, routeAgentRequest } from "agents";
 import type {
   FiberContext,
   FiberRecoveryContext,
-  FiberInspection,
   FiberRecoveryResult,
   SubAgentStub
 } from "agents";
 import { Chat } from "chat";
 import type { Message, Thread } from "chat";
+import {
+  adminConversationFromRow,
+  adminReplyJobFromFiber,
+  conversationTitle,
+  messagePreview,
+  providerFromThreadId,
+  type AdminConversation,
+  type AdminConversationRow,
+  type AdminReplyJob,
+  type AdminSetupInfo
+} from "./admin/directory";
 import { APPROVE_ACTION_ID, REJECT_ACTION_ID } from "./demos";
 import { ConversationAgent } from "./intelligence/conversation-agent";
+import {
+  AI_REPLY_FIBER_NAME,
+  EMPTY_AI_RESPONSE,
+  INTERRUPTED_AI_RESPONSE,
+  aiReplyFailureMode,
+  aiReplyRecoveryMode,
+  aiReplySnapshot,
+  parseAiReplySnapshot,
+  type AiReplySnapshot
+} from "./intelligence/delivery";
 import {
   conversationNameForThread,
   isMenuCommand,
@@ -27,192 +47,31 @@ import {
   postMainMenu,
   postMenu
 } from "./menu";
+import {
+  TELEGRAM_STREAM_SOFT_LIMIT,
+  WEBHOOK_PATH,
+  isExpectedFinalEditNoop,
+  setupTelegramWebhook,
+  shardTelegramStateKey,
+  splitTelegramMessageText
+} from "./provider/telegram";
 import { createAgentChatState } from "./state";
 
 export { ConversationAgent } from "./intelligence/conversation-agent";
 export { ChatStateAgent } from "./state";
 
-const WEBHOOK_PATH = "/webhooks/telegram";
+export type {
+  AdminConversation,
+  AdminReplyJob,
+  AdminSetupInfo
+} from "./admin/directory";
+export type { AiReplySnapshot } from "./intelligence/delivery";
+export type { TelegramWebhookSetupResult } from "./provider/telegram";
+
 const DEFAULT_AGENT_NAME = "default";
-const AI_REPLY_FIBER_NAME = "chat-sdk-messenger:ai-reply";
-const TELEGRAM_API_BASE = "https://api.telegram.org";
-const TELEGRAM_STREAM_SOFT_LIMIT = 3_400;
-const TELEGRAM_FOLLOWUP_CHUNK_LIMIT = 3_500;
-const EMPTY_AI_RESPONSE =
-  "I couldn't produce a text response. Please try again.";
-const INTERRUPTED_AI_RESPONSE =
-  "Sorry, my reply was interrupted. Please send your message again if you'd like me to retry.";
-
-export type AiReplyStage = "accepted" | "streaming" | "completed";
-
-export type AiReplySnapshot = {
-  type: typeof AI_REPLY_FIBER_NAME;
-  stage: AiReplyStage;
-  thread: unknown;
-  message: unknown;
-};
-
-export type AdminConversation = {
-  threadId: string;
-  conversationName: string;
-  provider: string;
-  title: string;
-  lastMessagePreview?: string;
-  createdAt: number;
-  lastMessageAt: number;
-};
-
-export type AdminReplyJob = {
-  fiberId: string;
-  status: FiberInspection["status"];
-  threadId?: string;
-  messageId?: string;
-  createdAt: number;
-  startedAt?: number;
-  settledAt?: number;
-  error?: string;
-};
-
-export type AdminSetupInfo = {
-  webhookPath: string;
-  agentName: string;
-  telegramConfigured: boolean;
-  telegramUserName: string;
-};
-
-export type TelegramWebhookSetupResult = {
-  ok: boolean;
-  webhookUrl: string;
-  alreadyConfigured: boolean;
-  description: string;
-};
-
-export function aiReplyRecoveryMode(
-  snapshot: AiReplySnapshot
-): "answer" | "apologize" | null {
-  if (snapshot.stage === "accepted") {
-    return "answer";
-  }
-  if (snapshot.stage === "streaming") {
-    return "apologize";
-  }
-  return null;
-}
-
-export function aiReplyFailureMode(
-  hasStreamedText: boolean,
-  completedModelTurn = false,
-  expectedDeliveryCompletion = false
-): "apologize" | "error" | null {
-  if (expectedDeliveryCompletion) {
-    return null;
-  }
-
-  if (completedModelTurn) {
-    return "error";
-  }
-
-  return hasStreamedText ? "apologize" : "error";
-}
-
-export function isIgnorableDeliveryError(error: unknown): boolean {
-  if (error === undefined || error === null) {
-    return false;
-  }
-
-  const candidate = error as { code?: unknown; message?: unknown };
-  const code = typeof candidate.code === "string" ? candidate.code : undefined;
-  const message =
-    typeof candidate.message === "string"
-      ? candidate.message.toLowerCase()
-      : String(error).toLowerCase();
-
-  return (
-    code === "VALIDATION_ERROR" && message.includes("message is not modified")
-  );
-}
-
-export function isExpectedFinalEditNoop(
-  error: unknown,
-  callback: Pick<TextStreamCallback, "visibleLimitReached">
-): boolean {
-  return callback.visibleLimitReached() && isIgnorableDeliveryError(error);
-}
-
-export function splitTelegramMessageText(
-  text: string,
-  limit = TELEGRAM_FOLLOWUP_CHUNK_LIMIT
-): string[] {
-  if (!text.trim()) {
-    return [];
-  }
-
-  const chunks: string[] = [];
-  let remaining = text;
-
-  while (remaining.length > limit) {
-    let splitAt = remaining.lastIndexOf("\n\n", limit);
-    if (splitAt < Math.floor(limit * 0.5)) {
-      splitAt = remaining.lastIndexOf("\n", limit);
-    }
-    if (splitAt < Math.floor(limit * 0.5)) {
-      splitAt = remaining.lastIndexOf(" ", limit);
-    }
-    if (splitAt < Math.floor(limit * 0.5)) {
-      splitAt = limit;
-    }
-
-    chunks.push(remaining.slice(0, splitAt));
-    remaining = remaining.slice(splitAt);
-  }
-
-  if (remaining) {
-    chunks.push(remaining);
-  }
-
-  return chunks;
-}
 
 function toError(error: unknown): Error {
   return error instanceof Error ? error : new Error(String(error));
-}
-
-function parseAiReplySnapshot(snapshot: unknown): AiReplySnapshot | null {
-  if (snapshot === null || typeof snapshot !== "object") {
-    return null;
-  }
-
-  const candidate = snapshot as Partial<AiReplySnapshot>;
-  if (
-    candidate.type !== AI_REPLY_FIBER_NAME ||
-    (candidate.stage !== "accepted" &&
-      candidate.stage !== "streaming" &&
-      candidate.stage !== "completed") ||
-    candidate.thread === undefined ||
-    candidate.message === undefined
-  ) {
-    return null;
-  }
-
-  return {
-    type: AI_REPLY_FIBER_NAME,
-    stage: candidate.stage,
-    thread: candidate.thread,
-    message: candidate.message
-  };
-}
-
-function aiReplySnapshot(
-  stage: AiReplyStage,
-  thread: Thread,
-  message: Message
-): AiReplySnapshot {
-  return {
-    type: AI_REPLY_FIBER_NAME,
-    stage,
-    thread: thread.toJSON(),
-    message: message.toJSON()
-  };
 }
 
 function setupErrorResponse(error: Error): Response {
@@ -227,169 +86,8 @@ function setupErrorResponse(error: Error): Response {
   );
 }
 
-function jsonResponse(value: unknown, init?: ResponseInit): Response {
-  return new Response(JSON.stringify(value, null, 2), {
-    ...init,
-    headers: {
-      "content-type": "application/json; charset=utf-8",
-      ...init?.headers
-    }
-  });
-}
-
-type TelegramApiResponse<T> =
-  | {
-      ok: true;
-      result: T;
-      description?: string;
-    }
-  | {
-      ok: false;
-      description?: string;
-      error_code?: number;
-    };
-
-type TelegramWebhookInfo = {
-  url?: string;
-};
-
-async function telegramApi<T>(
-  token: string,
-  method: string,
-  body?: Record<string, unknown>
-): Promise<TelegramApiResponse<T>> {
-  const response = await fetch(`${TELEGRAM_API_BASE}/bot${token}/${method}`, {
-    method: "POST",
-    headers: {
-      "content-type": "application/json"
-    },
-    body: JSON.stringify(body ?? {})
-  });
-  return (await response.json()) as TelegramApiResponse<T>;
-}
-
-async function setupTelegramWebhook(
-  request: Request,
-  env: Cloudflare.Env
-): Promise<Response> {
-  const requestUrl = new URL(request.url);
-  const webhookUrl = `${requestUrl.origin}${WEBHOOK_PATH}`;
-  if (requestUrl.protocol !== "https:") {
-    return jsonResponse(
-      {
-        ok: false,
-        webhookUrl,
-        error:
-          "Telegram webhooks require HTTPS. Open the Quick Tunnel or deployed Worker URL and click Set webhook here again."
-      },
-      { status: 400 }
-    );
-  }
-
-  if (!env.TELEGRAM_BOT_TOKEN) {
-    return jsonResponse(
-      {
-        ok: false,
-        error: "TELEGRAM_BOT_TOKEN is not configured."
-      },
-      { status: 400 }
-    );
-  }
-
-  if (!env.TELEGRAM_WEBHOOK_SECRET_TOKEN) {
-    return jsonResponse(
-      {
-        ok: false,
-        error: "TELEGRAM_WEBHOOK_SECRET_TOKEN is not configured."
-      },
-      { status: 400 }
-    );
-  }
-
-  const current = await telegramApi<TelegramWebhookInfo>(
-    env.TELEGRAM_BOT_TOKEN,
-    "getWebhookInfo"
-  );
-
-  if (!current.ok) {
-    return jsonResponse(
-      {
-        ok: false,
-        webhookUrl,
-        error: current.description ?? "Failed to inspect Telegram webhook."
-      },
-      { status: 502 }
-    );
-  }
-
-  if (current.result.url === webhookUrl) {
-    return jsonResponse({
-      ok: true,
-      webhookUrl,
-      alreadyConfigured: true,
-      description: "Telegram webhook already points at this origin."
-    } satisfies TelegramWebhookSetupResult);
-  }
-
-  const next = await telegramApi<true>(env.TELEGRAM_BOT_TOKEN, "setWebhook", {
-    url: webhookUrl,
-    secret_token: env.TELEGRAM_WEBHOOK_SECRET_TOKEN
-  });
-
-  if (!next.ok) {
-    return jsonResponse(
-      {
-        ok: false,
-        webhookUrl,
-        error: next.description ?? "Failed to set Telegram webhook."
-      },
-      { status: 502 }
-    );
-  }
-
-  return jsonResponse({
-    ok: true,
-    webhookUrl,
-    alreadyConfigured: false,
-    description: next.description ?? "Telegram webhook configured."
-  } satisfies TelegramWebhookSetupResult);
-}
-
 export function getIngressAgentName(_request: Request): string {
   return DEFAULT_AGENT_NAME;
-}
-
-function providerFromThreadId(threadId: string): string {
-  return threadId.split(":")[0] || "unknown";
-}
-
-function conversationTitle(thread: Thread): string {
-  return `${providerFromThreadId(thread.id)}:${thread.id.split(":")[1] ?? thread.id}`;
-}
-
-function messagePreview(message: Message): string {
-  return message.text.trim().slice(0, 160);
-}
-
-function metadataString(
-  metadata: Record<string, unknown> | null | undefined,
-  key: string
-): string | undefined {
-  const value = metadata?.[key];
-  return typeof value === "string" ? value : undefined;
-}
-
-function adminReplyJobFromFiber(fiber: FiberInspection): AdminReplyJob {
-  return {
-    fiberId: fiber.fiberId,
-    status: fiber.status,
-    threadId: metadataString(fiber.metadata, "threadId"),
-    messageId: metadataString(fiber.metadata, "messageId"),
-    createdAt: fiber.createdAt,
-    startedAt: fiber.startedAt,
-    settledAt: fiber.settledAt,
-    error: fiber.error
-  };
 }
 
 export class ChatIngressAgent extends Agent {
@@ -405,6 +103,111 @@ export class ChatIngressAgent extends Agent {
       this.bot = undefined;
       this.botStartupError = toError(error);
     }
+  }
+
+  private createBot(): Chat {
+    if (!this.env.TELEGRAM_BOT_TOKEN) {
+      throw new Error("TELEGRAM_BOT_TOKEN is required");
+    }
+
+    const userName =
+      this.env.TELEGRAM_BOT_USERNAME ?? "cloudflare_chat_sdk_bot";
+    const telegram = createTelegramAdapter({
+      botToken: this.env.TELEGRAM_BOT_TOKEN,
+      mode: "webhook",
+      secretToken: this.env.TELEGRAM_WEBHOOK_SECRET_TOKEN,
+      userName
+    });
+
+    const bot = new Chat({
+      userName,
+      adapters: { telegram },
+      state: createAgentChatState({
+        parent: this,
+        keyShard: (key) => shardTelegramStateKey(key, this.shardThread),
+        shardKey: this.shardThread
+      }),
+      concurrency: { strategy: "burst", debounceMs: 600 }
+    });
+
+    bot.onNewMention(async (thread, message) => {
+      await thread.subscribe();
+      if (isMenuCommand(message.text)) {
+        await postMainMenu(thread);
+        return;
+      }
+
+      await this.enqueueConversationReply(thread, message);
+    });
+
+    bot.onDirectMessage(async (thread, message) => {
+      if (isMenuCommand(message.text)) {
+        await postMainMenu(thread);
+        return;
+      }
+
+      if (isResetCommand(message.text)) {
+        await this.resetConversation(thread);
+        return;
+      }
+
+      await this.enqueueConversationReply(thread, message);
+    });
+
+    bot.onSubscribedMessage(async (thread, message) => {
+      if (isMenuCommand(message.text)) {
+        await postMainMenu(thread);
+        return;
+      }
+
+      if (isResetCommand(message.text)) {
+        await this.resetConversation(thread);
+        return;
+      }
+
+      if (this.shouldUseAi(message, thread)) {
+        await this.enqueueConversationReply(thread, message);
+      }
+    });
+
+    bot.onAction(async (event) => {
+      const thread = event.thread;
+      if (!thread) {
+        return;
+      }
+
+      if (event.actionId === ASK_AGENT_ACTION_ID) {
+        await postAskAgentInstructions(thread);
+        return;
+      }
+
+      if (MENU_IDS.has(event.actionId)) {
+        await postMenu(thread, event.actionId);
+        return;
+      }
+
+      const demo = DEMO_LOOKUP.get(event.actionId);
+      if (demo) {
+        await demo.run(thread);
+        return;
+      }
+
+      if (
+        event.actionId === APPROVE_ACTION_ID ||
+        event.actionId === REJECT_ACTION_ID
+      ) {
+        const decision =
+          event.actionId === APPROVE_ACTION_ID ? "approved" : "rejected";
+        await event.adapter.editMessage(event.threadId, event.messageId, {
+          markdown: `Deploy preview ${decision} by ${event.user.fullName || event.user.userName}.`
+        });
+        return;
+      }
+
+      await thread.post(`Unknown action: ${event.actionId}`);
+    });
+
+    return bot.registerSingleton();
   }
 
   private ensureAdminSchema(): void {
@@ -562,15 +365,7 @@ export class ChatIngressAgent extends Agent {
   }
 
   private readConversation(threadId: string): AdminConversation | null {
-    const rows = this.sql<{
-      thread_id: string;
-      conversation_name: string;
-      provider: string;
-      title: string;
-      last_message_preview: string | null;
-      created_at: number;
-      last_message_at: number;
-    }>`
+    const rows = this.sql<AdminConversationRow>`
       SELECT thread_id, conversation_name, provider, title,
              last_message_preview, created_at, last_message_at
       FROM chat_admin_conversations
@@ -582,27 +377,11 @@ export class ChatIngressAgent extends Agent {
       return null;
     }
 
-    return {
-      threadId: row.thread_id,
-      conversationName: row.conversation_name,
-      provider: row.provider,
-      title: row.title,
-      lastMessagePreview: row.last_message_preview ?? undefined,
-      createdAt: row.created_at,
-      lastMessageAt: row.last_message_at
-    };
+    return adminConversationFromRow(row);
   }
 
   private readConversations(): AdminConversation[] {
-    const rows = this.sql<{
-      thread_id: string;
-      conversation_name: string;
-      provider: string;
-      title: string;
-      last_message_preview: string | null;
-      created_at: number;
-      last_message_at: number;
-    }>`
+    const rows = this.sql<AdminConversationRow>`
       SELECT thread_id, conversation_name, provider, title,
              last_message_preview, created_at, last_message_at
       FROM chat_admin_conversations
@@ -610,15 +389,7 @@ export class ChatIngressAgent extends Agent {
       LIMIT 100
     `;
 
-    return rows.map((row) => ({
-      threadId: row.thread_id,
-      conversationName: row.conversation_name,
-      provider: row.provider,
-      title: row.title,
-      lastMessagePreview: row.last_message_preview ?? undefined,
-      createdAt: row.created_at,
-      lastMessageAt: row.last_message_at
-    }));
+    return rows.map(adminConversationFromRow);
   }
 
   private async recordConversation(
@@ -656,110 +427,6 @@ export class ChatIngressAgent extends Agent {
       createdAt: now,
       lastMessageAt: now
     };
-  }
-
-  private createBot(): Chat {
-    if (!this.env.TELEGRAM_BOT_TOKEN) {
-      throw new Error("TELEGRAM_BOT_TOKEN is required");
-    }
-
-    const userName =
-      this.env.TELEGRAM_BOT_USERNAME ?? "cloudflare_chat_sdk_bot";
-    const telegram = createTelegramAdapter({
-      botToken: this.env.TELEGRAM_BOT_TOKEN,
-      mode: "webhook",
-      secretToken: this.env.TELEGRAM_WEBHOOK_SECRET_TOKEN,
-      userName
-    });
-
-    const bot = new Chat({
-      userName,
-      adapters: { telegram },
-      state: createAgentChatState({
-        parent: this,
-        shardKey: (threadId) => threadId.split(":").slice(0, 2).join(":")
-      }),
-      concurrency: { strategy: "burst", debounceMs: 600 }
-    });
-
-    bot.onNewMention(async (thread, message) => {
-      await thread.subscribe();
-      if (isMenuCommand(message.text)) {
-        await postMainMenu(thread);
-        return;
-      }
-
-      await this.enqueueConversationReply(thread, message);
-    });
-
-    bot.onDirectMessage(async (thread, message) => {
-      if (isMenuCommand(message.text)) {
-        await postMainMenu(thread);
-        return;
-      }
-
-      if (isResetCommand(message.text)) {
-        await this.resetConversation(thread);
-        return;
-      }
-
-      await this.enqueueConversationReply(thread, message);
-    });
-
-    bot.onSubscribedMessage(async (thread, message) => {
-      if (isMenuCommand(message.text)) {
-        await postMainMenu(thread);
-        return;
-      }
-
-      if (isResetCommand(message.text)) {
-        await this.resetConversation(thread);
-        return;
-      }
-
-      if (this.shouldUseAi(message, thread)) {
-        await this.enqueueConversationReply(thread, message);
-      }
-    });
-
-    bot.onAction(async (event) => {
-      const thread = event.thread;
-      if (!thread) {
-        return;
-      }
-
-      if (event.actionId === ASK_AGENT_ACTION_ID) {
-        await postAskAgentInstructions(thread);
-        return;
-      }
-
-      if (MENU_IDS.has(event.actionId)) {
-        await postMenu(thread, event.actionId);
-        return;
-      }
-
-      const demo = DEMO_LOOKUP.get(event.actionId);
-      if (demo) {
-        await demo.run(thread);
-        return;
-      }
-
-      if (
-        event.actionId === APPROVE_ACTION_ID ||
-        event.actionId === REJECT_ACTION_ID
-      ) {
-        const decision =
-          event.actionId === APPROVE_ACTION_ID ? "approved" : "rejected";
-        await event.adapter.editMessage(event.threadId, event.messageId, {
-          markdown: `Deploy preview ${decision} by ${event.user.fullName || event.user.userName}.`
-        });
-        return;
-      }
-
-      await thread.post(`Unknown action: ${event.actionId}`);
-    });
-
-    return bot.registerSingleton();
   }
 
   private async answerWithConversationAgent(
@@ -876,6 +543,10 @@ export class ChatIngressAgent extends Agent {
     thread: Thread
   ): Promise<SubAgentStub<ConversationAgent>> {
     return this.subAgent(ConversationAgent, conversationNameForThread(thread));
+  }
+
+  private shardThread(threadId: string): string {
+    return threadId.split(":").slice(0, 2).join(":");
   }
 
   private shouldUseAi(message: Message, thread: Thread): boolean {
